requirements.txt
# ============================================================
# SAM Assistive Vision — Python Dependencies
# ============================================================
# Install with: pip install -r requirements.txt
# ============================================================

# --- Web Framework & Server ---
fastapi==0.115.5
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.12

# --- Configuration ---
pydantic-settings==2.6.1

# --- Image Processing ---
pillow>=11.1.0
numpy>=2.1.0
scipy>=1.14.0

# --- HTTP Client (Ollama API calls) ---
httpx==0.27.2

# --- TTS (Stage 6 -- active) ---
edge-tts>=7.0.0

# --- STT (Stage 1 -- active) ---
faster-whisper>=1.0.0
soundfile>=0.12.0
sounddevice>=0.4.0        # Microphone recording (tests/mic_test.py demo)

# --- Vision Models (Stages 3-4 -- active) ---
# Install torch/torchvision separately for your target runtime.
# Local CUDA setup uses the cu128 wheel index before installing this file.
ultralytics>=8.0.0
transformers>=4.40.0
accelerate>=0.27.0

# --- OCR (Stage 4b -- active in OCR mode) ---
paddlepaddle==3.0.0
paddleocr==3.4.0
